answer/Question_3/ruyi-url-checker/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
httpx>=0.25.2
python-dotenv>=1.0.0
toml>=0.10.2
gitpython>=3.1.40
apscheduler>=3.10.4
pytz>=2023.3
loguru>=0.7.2
